requirements.txt
Flask>=3.1
psycopg2-binary>=2.9
requests>=2.32
